Changed grid size

+++ games/TurnBasedGames/Bouncy/js/GameElements/Projectiles/Projectile.js
@@ -150,7 +150,12 @@ class Projectile {
 
     this.getStyle().rotateProjectile(this, this.gameSprite);
 
-    if (this.x <= 0 || this.x > boardWidth || this.y < 0 || this.y > boardHeight) {
+    if (
+      this.x <= 0 && !(this.gravity && this.gravity.x > 0) || 
+      this.x > boardWidth && !(this.gravity && this.gravity.x < 0) || 
+      this.y < 0 && !(this.gravity && this.gravity.y > 0) || 
+      this.y > boardHeight && !(this.gravity && this.gravity.y < 0)
+    ) {
       this.delete();
     }
   }
@@ -167,7 +172,18 @@ class Projectile {
   }
 
   hitWall(boardState, intersection) {
-    if (this.destroyOnWall) { this.delete(); }
+    if (this.destroyOnWall) { 
+      if (intersection.line instanceof BorderWallLine) {
+        if (!(
+          intersection.line.side == BorderWallLine.LEFT && this.gravity && this.gravity.x > 0 ||
+          intersection.line.side == BorderWallLine.RIGHT && this.gravity && this.gravity.x < 0 ||
+          intersection.line.side == BorderWallLine.TOP && this.gravity && this.gravity.y > 0 ||
+          intersection.line.side == BorderWallLine.BOTTOM && this.gravity && this.gravity.y < 0
+        )) {
+          this.delete();
+        }
+      }
+    }
     this.wallsHit += 1;
   }
